requirements.txt
pandas>=1.3.0
numpy>=1.21.0
pyyaml>=6.0
tqdm>=4.60.0